Move JobTemplate model to a base_models directory
https://bugzilla.redhat.com/show_bug.cgi?id=1369842
