+++ lib/ansible_tower_client/base_models/job_template.rb
